savedcharacters file added
